src: avoid silent coercion to signed/unsigned int
Be accurate about signedness and bit widths.

PR-URL: https://github.com/nodejs/node/pull/50663

+++ src/permission/fs_permission.cc
@@ -56,7 +56,7 @@ bool is_tree_granted(node::permission::FSPermission::RadixTree* granted_tree,
   // is UNC file path
   if (param.rfind("\\\\", 0) == 0) {
     // return lookup with normalized param
-    int starting_pos = 4;  // "\\?\"
+    size_t starting_pos = 4;  // "\\?\"
     if (param.rfind("\\\\?\\UNC\\") == 0) {
       starting_pos += 4;  // "UNC\"
     }
@@ -176,7 +176,7 @@ bool FSPermission::RadixTree::Lookup(const std::string_view& s,
   if (current_node->children.size() == 0) {
     return when_empty_return;
   }
-  unsigned int parent_node_prefix_len = current_node->prefix.length();
+  size_t parent_node_prefix_len = current_node->prefix.length();
   const std::string path(s);
   auto path_len = path.length();
 
@@ -202,10 +202,10 @@ bool FSPermission::RadixTree::Lookup(const std::string_view& s,
 void FSPermission::RadixTree::Insert(const std::string& path) {
   FSPermission::RadixTree::Node* current_node = root_node_;
 
-  unsigned int parent_node_prefix_len = current_node->prefix.length();
-  int path_len = path.length();
+  size_t parent_node_prefix_len = current_node->prefix.length();
+  size_t path_len = path.length();
 
-  for (int i = 1; i <= path_len; ++i) {
+  for (size_t i = 1; i <= path_len; ++i) {
     bool is_wildcard_node = path[i - 1] == '*';
     bool is_last_char = i == path_len;
